requirements.txt
absl-py
aiohttp
argon2-cffi
astunparse==1.6.3
async-generator
async-timeout==3.0.1
attrs
backcall
bleach
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
catboost==0.25.1
certifi==2021.5.30
cffi
chardet==4.0.0
click
cloudpickle
colorama
coverage
cryptography
cycler==0.10.0
Cython
cytoolz==0.11.0
dask
decorator
defusedxml
dlib
entrypoints==0.3
filelock==3.0.12
Flask
flatbuffers==1.12
fsspec
gast==0.3.3
gdown==3.13.0
google-auth==1.30.0
google-auth-oauthlib
google-pasta==0.2.0
graphviz==0.16
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imagecodecs
imageio
imageio-ffmpeg==0.4.3
importlib-metadata
ipykernel
ipython
ipython-genutils
itsdangerous
jedi==0.17.0
Jinja2
joblib
jsonschema
jupyter-client
jupyter-core
jupyterlab-pygments
Keras-Applications
Keras-Preprocessing
kiwisolver
llvmlite==0.34.0
locket==0.2.1
Markdown
MarkupSafe==1.1.1
matplotlib
mediapipe==0.8.4.2
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
moviepy==1.0.3
multidict
natsort
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
numba==0.51.2
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.2.52
opt-einsum==3.3.0
packaging
pandas
pandocfilters
parso
partd
pickleshare
Pillow
plotly==4.14.3
proglog==0.1.9
prometheus-client
prompt-toolkit
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
Pygments
PyJWT==1.7.1
pyOpenSSL
pyparsing
pyreadline==2.1
pyrsistent
PySocks
python-dateutil
pytz
PyWavelets
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
sacrebleu==1.5.1
scikit-image==0.18.1
scikit-learn
scipy
seaborn
Send2Trash
sip==4.19.13
six==1.15.0
tensorboard
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.4
testpath
threadpoolctl
tifffile
toolz
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado
tqdm==4.60.0
traitlets
typing-extensions
urllib3==1.26.4
wcwidth
webencodings==0.5.1
Werkzeug==1.0.1
win-inet-pton
wincertstore==0.2
wrapt==1.12.1
yarl
zipp
